% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-contrast.R
\name{aes_contrast}
\alias{aes_contrast}
\title{Modify an aesthetic for contrast}
\usage{
aes_contrast(..., dark = NULL, light = NULL, aesthetic = "colour")
}
\arguments{
\item{...}{Require named arguments (and support trailing commas).}

\item{dark}{A dark colour. If NULL, derived from theme text or panel background.}

\item{light}{A light colour. If NULL, derived from theme text or panel background.}

\item{aesthetic}{The aesthetic to be modified for contrast. Either "colour" (default) or "fill".}
}
\value{
A ggplot2 aesthetic in \link[ggplot2:aes]{ggplot2::aes}.
}
\description{
Modifies a mapped colour (or fill) aesthetic for contrast against the fill (or colour) aesthetic.

Function can be spliced into \link[ggplot2:aes]{ggplot2::aes} with \link[rlang:splice-operator]{rlang::!!!}.
}
\examples{
library(ggplot2)
library(dplyr)
library(stringr)

palmerpenguins::penguins |>
  count(species, sex) |>
  mutate(sex = str_to_sentence(sex)) |>
  ggplot(aes(x = sex, y = n, fill = species, label = n)) +
  geom_col(
    position = position_dodge(preserve = "single"),
    width = 0.75,
  ) +
  scale_fill_discrete(palette = c("navy", "orange", "pink")) +
  geom_text(
    mapping = aes_contrast(),
    position = position_dodge(width = 0.75, preserve = "single"),
    vjust = 1.33,
    show.legend = FALSE,
  )

palmerpenguins::penguins |>
  count(species, sex) |>
  mutate(sex = str_to_sentence(sex)) |>
  ggplot(aes(y = sex, x = n, fill = species, label = n)) +
  geom_col(
    position = position_dodge(preserve = "single"),
    width = 0.75,
  ) +
  scale_fill_discrete(palette = c("navy", "orange", "pink")) +
  geom_text(
    mapping = aes_contrast(),
    position = position_dodge(width = 0.75, preserve = "single"),
    hjust = 1.25,
    show.legend = FALSE,
  )

}
\seealso{
\code{\link[rlang]{splice}}
}
