% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-each-all.R
\name{bind_each_all}
\alias{bind_each_all}
\title{Bind each all.}
\usage{
bind_each_all(
  data,
  ...,
  name = "each_all",
  each = "Each",
  all = "All",
  after = Inf
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{...}{Require named arguments (and support trailing commas).}

\item{name}{A variable name. Defaults to \code{"each_all"}.}

\item{each}{A string for the each value. Defaults to \code{"Each"}.}

\item{all}{A string for the all value. Defaults to \code{"All"}.}

\item{after}{A number for where the all value should be placed after. Defaults to \code{Inf}, which puts \code{"All"} last. Use \code{0} to put \code{"All"} first.}
}
\value{
A data frame or tibble
}
\description{
Binds data to support plotting each category \emph{and} all combined data.
}
\examples{
library(dplyr)
library(ggplot2)

palmerpenguins::penguins |>
  count(species)

palmerpenguins::penguins |>
  bind_each_all(species) |>
  count(species, each_all)

palmerpenguins::penguins |>
  bind_each_all(species) |>
  ggplot(aes(x = species, y = body_mass_g)) +
  geom_jitter()

palmerpenguins::penguins |>
  bind_each_all(species, all = "All\nspecies") |>
  ggplot(aes(x = species, y = body_mass_g, colour = each_all)) +
  geom_jitter() +
  facet_wrap(facets = vars(each_all), scales = "free_x", space = "free_x") +
  theme(strip.text.x = element_blank()) +
  theme(legend.position = "none") +
  labs(x = NULL) +
  scale_colour_discrete(palette = c("steelblue", "darkgrey"))

}
