% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggetc-package.R
\docType{package}
\name{ggetc-package}
\alias{ggetc}
\alias{ggetc-package}
\title{ggetc: Assorted 'ggplot2' Helper Functions}
\description{
Assorted 'ggplot2' helper functions.
}
\author{
\strong{Maintainer}: David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID}) [copyright holder]

}
\keyword{internal}
